Implement batch mode
Implement a new mode which asks imx_usb to explicitly work on a
specific device PID/VID combination. This mode uses imx_usb.conf not
as a list of possible devices, but as a list of devices we must work
through. It can be used with a single device only (making sure that
only the device specified in the configuration file gets handled) or
through multiple devices with different PID/VID. The latter is mainly
ment for devices which reenumerate with a different PID/VID. The mode
allows to download the SPL using Boot ROM first, and then U-Boot using
SPL's Serial Download Protocol (SDP) emulation next.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>